requirements.txt
asgiref==3.9.1
bleach==6.2.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coverage==7.10.6
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.5
django-allauth==65.10.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
et_xmlfile==2.0.0
idna==3.10
numpy==2.3.1
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
